rewrite parser from scratch
the initial methodology for parsing was flawed. this is a complete
rewrite from the bottom up (i.e. starting with terms then moving onto
operators)

as of this commit, constants and variables are supported, as well as
parenthetical expressions (containing only constants and variables) and
concatenation as a multiplication operator

e.g. "3(3x)" will parse successfully