add separate threads for input and ui
in this model, the application state is shared between threads. when
some event happens (like a ui update or an input), the thread waits for
a lock on the application state. it may be better to switch to a
message based approach, but then much of the processing still happens
on the main thread (which might be okay)